Create doc during token save
We needed to do this because otherwise it throws an
error complaining about not having a document to
update. It will still update the document if it exists,
though, so the functionality should be the same.

This also includes an extra test for the NotificationService,
which should bring test coverage to 100% for that file.